requirements.txt
PyQt5>=5.15.0
Pillow>=8.0.0
imageio>=2.9.0
imageio-ffmpeg>=0.4.0
pyinstaller>=5.0.0